Invoice work
Edited the EmployeePage's "view sold/leased cars" funciton to better represent those cars (added vins, took out information dependent on invoices.
Also added two new classes to grant access to invocies. The idea is that the employee will first go to the InoviceLookupPage, which is very similar to the DealershipLookupPage (it contains rows of all invoices and a search field to search by a specific invoice ID). Upon typing in a specific ID, a new page would appear displaying the information in a format similar to the view we had in our brainstorming (work in progress, so far i've been organizing variables and ways to access them from the database, no GUI yet.